requirements.txt
blinker==1.9.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
loguru==0.7.2
MarkupSafe==3.0.2
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
pathlib==1.0.1
pillow==11.0.0
pyee==12.1.1
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pysubs2==1.7.3
python-dateutil==2.9.0.post0
python-ffmpeg==2.0.12
six==1.16.0
sub2clip @ git+https://github.com/lpalinckx/sub2clip.git@305270bc0f784a78c8085deb40ec72371f7ccb3d
typing_extensions==4.12.2